backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.38
asyncpg==0.31.0
pydantic==2.11.1
pydantic-settings==2.8.1
alembic==1.15.1
google-genai
python-dotenv==1.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
